Create an interactome from BioGRID database or input custom interactome

This function creates an interactome before conducting network analysis for Time-series data. Different from interactome which can either be performed before preprocess or after preprocess, this method must be performed after preprocessNwaTS.

Usage

interactomeNwaTS(
  nwaList,
  interactionMatrix = NULL,
  species,
  link = defaultBioGridLink(),
  reportDir = "HTSanalyzerReport",
  genetic = FALSE,
  force = FALSE,
  verbose = TRUE
)

Arguments

nwaList

A named list of NWA object generated by preprocessNwaTS.

interactionMatrix

An interaction matrix including columns 'InteractionType', 'InteractorA' and 'InteractorB'. If this matrix is available, the interactome can be directly built based on it instead of downloading from BioGRID.

species

A single character value specifying the species for which the data should be read.

link

the link (url) where the data should be downloaded (in tab2 format). The default link is version 3.4.138 of BioGRID.

reportDir

A single character value specifying the directory to store reports. The BioGRID data set will be downloaded and stored in a subdirectory called 'Data' in 'reportDir'.

genetic

A single logical value. If TRUE, genetic interactions will be kept; otherwise, they will be removed from the data set.

force

Force to download the data set.

verbose

A single logical value indicating to display detailed messages (when verbose=TRUE) or not (when verbose=FALSE).

Value

In the end, this function will return an updated list of NWA object.

See also

Examples

data(d7, d13, d25)

## generate expInfor to describe the information of time series data
expInfor <- matrix(c("d7", "d13", "d25"), nrow = 3, ncol = 2,
                   byrow = FALSE, dimnames = list(NULL, c("ID", "Description")))

## package pvalueTS into a list of pvalues
datalist <- list(d7, d13, d25)
pvalueTS <- lapply(datalist, function(x){
                   tmp <- as.vector(x$neg.p.value)
                   names(tmp) <- x$id
                   tmp})

## package phenotypeTS into a list of phenotypes if you want to color nodes by it,
## otherwise ignore it!
phenotypeTS <- lapply(datalist, function(x) {
                      tmp <- as.vector(x$neg.lfc)
                      names(tmp) <- x$id
                      tmp})

## Example1: create an object of class 'NWABatch' by inputting an igraph object as the interactome
data(Biogrid_HS_Interactome)
nwaTS <- NWABatch(expInfor = expInfor, pvalueTS = pvalueTS,
                 phenotypeTS = phenotypeTS, interactome = Biogrid_HS_Interactome)
#> This graph was created by an old(er) igraph version.
#> ℹ Call `igraph::upgrade_graph()` on it to use with the current igraph version.
#> For now we convert it on the fly...

## Example2: create an object of class 'NWABatch' without interactome
nwaTS <- NWABatch(expInfor = expInfor, pvalueTS = pvalueTS, phenotypeTS = phenotypeTS)
## preprocess nwaTS
nwaTS1 <- preprocessNwaTS(nwaTS, species="Hs", initialIDs="SYMBOL",
                         keepMultipleMappings=TRUE, duplicateRemoverMethod="max")
#> -Preprocessing for input p-values and phenotypes ...
#> --Removing invalid p-values and phenotypes ...
#> --Removing duplicated genes ...
#> --Converting annotations ...
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> -Preprocessing complete!
#> 
#> -Preprocessing for input p-values and phenotypes ...
#> --Removing invalid p-values and phenotypes ...
#> --Removing duplicated genes ...
#> --Converting annotations ...
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> -Preprocessing complete!
#> 
#> -Preprocessing for input p-values and phenotypes ...
#> --Removing invalid p-values and phenotypes ...
#> --Removing duplicated genes ...
#> --Converting annotations ...
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> 'select()' returned 1:many mapping between keys and columns
#> -- 602 genes (out of 8000) could not be mapped to any identifier, and were removed from the data. 
#> -Preprocessing complete!
#> 
## create an interactome for NWABatch by inputting an interaction matrix
data(Biogrid_HS_Mat)
nwaTS2 <- interactomeNwaTS(nwaTS1, interactionMatrix = Biogrid_HS_Mat, genetic=FALSE)
#> -Creating interactome ...
#> -Interactome created! 
#> 
#> -Creating interactome ...
#> -Interactome created! 
#> 
#> -Creating interactome ...
#> -Interactome created! 
#> 

if (FALSE) { # \dontrun{
## Example3: create an object of class 'NWABatch' without interactome
nwaTS <- NWABatch(expInfor = expInfor, pvalueTS = pvalueTS, phenotypeTS = phenotypeTS)
## preprocess nwaTS
nwaTS1 <- preprocessNwaTS(nwaTS, species="Hs", initialIDs="SYMBOL",
                         keepMultipleMappings=TRUE, duplicateRemoverMethod="max")
## create an interactome for nwa by downloading for BioGRID database
nwaTS2 <- interactomeNwaTS(nwaTS1, species="Hs", reportDir="HTSanalyzerReport", genetic=FALSE)
} # }